+++ app/src/main/java/com/edplan/framework/timing/AbstractClock.java
@@ -3,5 +3,5 @@ package com.edplan.framework.timing;
 
 public interface AbstractClock
 {
-	public int getTime();
+	public double getTime();
 }
