+++ app/src/main/java/com/edplan/framework/view/advance/widget/OsuTriangleField.java
@@ -25,7 +25,7 @@ public class OsuTriangleField extends BaseView
 	public void draw(Canvas canvas){
 		// TODO: Implement this method
 		super.draw(canvas);
-		triangleManager.measure(canvas,getContext().getFrameDeltaTime());
+		triangleManager.measure(canvas,(int)getContext().getFrameDeltaTime());
 		for(OsuTriangle t:triangleManager.getTriangles()){
 			t.drawOnCanvas(canvas);
 		}
