+++ app/src/main/java/com/edplan/nso/ruleset/std/playing/drawable/piece/ComboIndexPiece.java
@@ -74,10 +74,10 @@ public class ComboIndexPiece extends BasePiece
 		}
 
 		@Override
-		public void setProgressTime(int p) {
+		public void setProgressTime(double p) {
 			// TODO: Implement this method
 			super.setProgressTime(p);
-			float fp=p/(float)getDuration();
+			float fp=(float)(p/getDuration());
 			setAlpha(FMath.sin(fp*FMath.PiHalf));
 		}
 
@@ -100,10 +100,10 @@ public class ComboIndexPiece extends BasePiece
 		}
 
 		@Override
-		public void setProgressTime(int p) {
+		public void setProgressTime(double p) {
 			// TODO: Implement this method
 			super.setProgressTime(p);
-			float fp=p/(float)getDuration();
+			float fp=(float)(p/(float)getDuration());
 			setAlpha(1-fp);
 			float s=1.0f*(1-fp)+1.5f*fp;
 			setScale(s,s);
