+++ app/src/main/java/com/edplan/superutils/interfaces/AbstractLooper.java
@@ -2,7 +2,7 @@ package com.edplan.superutils.interfaces;
 
 public interface AbstractLooper<T extends Loopable>
 {
-	public void loop(int deltaTime);
+	public void loop(double deltaTime);
 	
 	public void prepare();
